requirements.txt
flask==2.0.1
jieba==0.42.1
python-dotenv==0.19.0
gunicorn==20.1.0